added marta's tree builder example
Here is a print out of all observation states in each maze world.

As a simple example, take a look at the testWorld.txt and treeBuilderTestWorld.csv.

world -- is the world name
nodeLocation - where in the world observation occurs
NID -- node ID
PARENTID - parent node ID
pathFromParent - path to arrive at this node
depth - how deep is the node in the tree
N - number of cells open at this node
S - steps from parent until this node
stepsFromRoot - the length of path from the starting location until this node
isLeaf - is this node a leaf
nodeValue - the value of this node (intermediate nodes show truncated values)

See if you can write a script to visulaize a graph based on this information. It will be easier if you start with the
testWorld, since some of the trees are quite big.