updated planningpolicy. shows probabilities only for locations with more than one choice
Here is the new planning policy file. It has only decision locations.

"actionstaken" shows how many of the directions were actually used at that location; So, when actionstaken = 1, that means that everyone went in the same direction. We have 138 decision data-points in total, from which we have 78 data-points such that actionstaken > 1.