added positional_nested_list to analyze_path()
positional_nested_list contains lists of the value choices available to the agent at each node. The values are positioned according to the direction in which the agent had to travel in order to reach the node with that terminal value. The order is Left, Right, Up, Down; this follows the order of the probabilities of going in a certain direction found in planning_policy.csv